Dumping what I have.
Nothing is rendering to the framebuffer in the deferred renderer.

The forward renderer should not work at all. Shaders aren't in the code ATM.

When the rendering starts to work again, the skeleton system would instantly work. The issue should not be with the bone transformations (which are all identity matrices).

Added a lot of random files because .gitignore never works for me.

I think I must have updated the SPMBuild tool chain. I would have to fix that as I removed a bunch of files to make it work (math).

I will continue to try and fix this just as an exercise so I can learn more C++ and eventually directx when I add skeletons and deferred rendering for that.